post-move-fixes: Moved misplaced files to the correct locations.
