+++ src/main/java/uk/ac/cam/cares/jps/base/query/KnowledgeBaseClient.java
