+++ src/main/java/uk/ac/cam/cares/jps/base/scenario/JenaReadHook.java
