+++ src/main/java/uk/ac/cam/cares/jps/base/slurm/job/configuration/SpringConfiguration.java
@@ -1,0 +1,16 @@
+package uk.ac.cam.cares.jps.base.slurm.job.configuration;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
+
+@Configuration
+@ComponentScan
+public class SpringConfiguration {
+
+  @Bean
+  public static PropertySourcesPlaceholderConfigurer propertyConfigInDev() {
+    return new PropertySourcesPlaceholderConfigurer();
+  }
+}
