+++ src/main/uk/ac/cam/cares/jps/base/query/JenaHelper.java
