+++ src/main/uk/ac/cam/cares/jps/base/query/sparql/JenaModelWrapper.java
