+++ src/main/uk/ac/cam/cares/jps/base/query/sparql/Prefixes.java
