+++ src/main/uk/ac/cam/cares/jps/base/util/CommandHelper.java
