+++ src/main/uk/ac/cam/cares/jps/base/util/MiscUtil.java
